// locations-frontend/src/api/endpoints/locations/locations.zod.ts
export const locationsControllerCreateBodyNameMin = 2;
export const locationsControllerCreateBodyNameMax = 60;

export const locationsControllerCreateBodyCoordinatesLonMin = -180;
export const locationsControllerCreateBodyCoordinatesLonMax = 180;

export const locationsControllerCreateBodyCoordinatesLatMin = -90;
export const locationsControllerCreateBodyCoordinatesLatMax = 90;

export const locationsControllerCreateBodyAddressMax = 120;

export const locationsControllerCreateBodyNotesMax = 500;



export const LocationsControllerCreateBody = zod.object({
  "name": zod.string().min(locationsControllerCreateBodyNameMin).max(locationsControllerCreateBodyNameMax),
  "category": zod.enum(['office', 'store', 'landmark']),
  "coordinates": zod.object({
  "lon": zod.number().min(locationsControllerCreateBodyCoordinatesLonMin).max(locationsControllerCreateBodyCoordinatesLonMax),
  "lat": zod.number().min(locationsControllerCreateBodyCoordinatesLatMin).max(locationsControllerCreateBodyCoordinatesLatMax)
}),
  "address": zod.string().max(locationsControllerCreateBodyAddressMax).optional(),
  "notes": zod.string().max(locationsControllerCreateBodyNotesMax).optional()
})